requirements.txt
# Dependências essenciais para AI-Cop Face Recognition API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Face Recognition
face-recognition>=1.3.0
Pillow>=10.1.0
dlib>=19.24.0
numpy>=1.24.0
scipy>=1.11.0

# Opcional para melhor performance
opencv-python>=4.8.0
